requirements.txt
fastapi
psycopg2
psycopg2-binary
requests
python-dotenv
uvicorn